requirements.txt
shap==0.40.0
scikit-learn==0.24.2
pandas
torch
tb-nightly
seaborn
future
h5py
wandb==0.12.6
google-cloud
google-cloud-storage
plotly~=5.3.1
pytz~=2021.3
setuptools~=58.5.3
numpy==1.20
requests
torchvision>=0.6.0
mpld3~=0.5
numba>=0.50
sphinx
sphinx-rtd-theme
sphinx-autodoc-typehints
